fix warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/module-pgsql/trunk@5085 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ src/QorePGConnection.cpp
@@ -173,7 +173,7 @@ qore_pg_numeric_out::qore_pg_numeric_out(const QoreNumberNode* n) : size(0) {
    }
    
    // now add digits after the decimal point
-   if (i != str.size()) {
+   if (i != (int)str.size()) {
       // skip decimal point
       ++i;
       di = str.size();
@@ -208,7 +208,7 @@ void qore_pg_numeric_out::convertToNet() {
 
    //printd(5, "qore_pg_numeric_out::convertToNet() ndigits: %hd weight: %hd sign: %hd dscale: %hd size: %d\n", ndigits, weight, sign, dscale, size);
    assert(ndigits < QORE_MAX_DIGITS);
-   for (unsigned i = 0; i < ndigits; ++i) {
+   for (unsigned i = 0; i < (unsigned)ndigits; ++i) {
       //printd(5, " + %hu\n", digits[i]);
       digits[i] = htons(digits[i]);
    }
